LayoutTests/platform/mac/pointer-lock/pointer-lock-option-unadjusted-movement-expected.txt

Test pointer lock unadjustedMovement option.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.testRunner is defined.
     requestPointerLock()
PASS [object Promise] is non-null.
     Promise resolved.
     requestPointerLock({unadjustedMovement: true})
PASS [object Promise] is non-null.
     unadjustedMovement is supported.
     Unlock
PASS onpointerlockchange received after: Unlock
     Lock with PointerLockOptionsEnabled = false
PASS undefined is undefined.
PASS onpointerlockchange received after: Lock with PointerLockOptionsEnabled = false
     Unlock
PASS onpointerlockchange received after: Unlock
PASS successfullyParsed is true

TEST COMPLETE
